Stop leaking a tokio::sync::broadcast::Receiver
Overall this cleans up the Stream API, consolidates concerns around
managing the ffmpeg process into one location, and removes the dangling
`Receiver`.

Previously, the `Stream` struct contained a copy of the receiving
channel, and was cloned for each active connection. This led to a
dangling instance of `Receiver`, which in turn causes the channel to
consistently lag, adding a delay roughly equal to the duration of the
buffer to all active streams.

Now, `Stream` is not used as the shared `State` for HTTP connections,
`Arc<Process>` is -- this allows handlers to take a clone of the
ref-counted `Process` and from that request a `Stream` containing the
`Receiver`.

This change also surfaced a flaw in the async frame reader loop --
previously, errors encountered when sending a frame across the broadcast
channel were treated as failures. In fact, a broadcast `Sender` will
only ever return an Error from `send` if there are no active
subscribers, in which case we actually want to drop the frame and move
on, effectively waiting until we have active clients before we send
frames.