Migrate from axum to tide
Tide is a framework-agnostic async web server, as opposed to axum which
explicitly builds on tokio.

In order to enable this migration, the server handler's abstraction of
the receive side of the broadcast channel needed to move from a `Stream`
to an `AsyncRead`. This was done via `TryStreamExt::into_async_read`,
which in turn required that the channel itself move from emitting a
`Frame` to emitting `io::Result<Frame>`, which in turn led to changes in
the errors associated with `Frame` operations.